Add dtrace support.
This patch is experimental and based off of work done here:

    http://bugs.ruby-lang.org/issues/2565

I've not isolated whitespace changed so expect the diff to be
relatively large as it currently stands. If official dtrace
doesn't land in trunk soon, I'll clean this up and make it easy
to apply to specific ruby releases as well. Enjoy!